pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-functions>=1.24.0",
    "bs4>=0.0.2",
    "flask>=3.1.3",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.9",
    "unstructured>=0.21.5",
]
